cmd/govim: tidy up; no code changes
* Define vimstate in its own file
* Move all hover stuff into hover.go
